Extracted cluster functions into files for responsibilities: election, leader, cluster
