+++ start.sh
@@ -1,1 +1,1 @@
-go run types.go cluster.go TestPersister.go raftstate.go raft.go
+go run debughelpers.go types.go election.go leader.go  cluster.go TestPersister.go raftstate.go raft.go
